Fixing memory editor because of changes to the ImGuiListClipper api.
Note that the demo is still a little buggy. It is very hard to select a cell by
mouse click. It only works if you click and release very fast. I think it is
because if the mouse release registers in a later frame then some other widget
is getting the event and the cell editor gets unfocused.